Refactor namespaces and update Enclosure entity
Updated namespaces from VirtualZooShared to VirtualZooAPI.
Moved ApplicationDbContext and factory classes to new namespaces.
Added SeedData class to initialize database with default data.
Updated EnclosureFactory to include zooId parameter.
Rounded SpaceRequirement and Size properties in factories.
Updated migration files and ApplicationDbContextModelSnapshot.
Added new migration to make ZooId non-nullable with default value.
Updated Program.cs and repository/service files for namespace change.
Added package reference to VirtualZooShared.csproj.
Updated Enclosure class to include ZooId property.
Removed old SeedData class from VirtualZooShared.Data.
